Change DataPoint time value to double

+++ app/src/main/java/ca/warp7/android/scouting/ScoutingActivity.kt
@@ -213,7 +213,7 @@ class ScoutingActivity : AppCompatActivity(), BaseScoutingActivity {
         })
         updateActivityStatus()
         updateCurrentTab()
-        entry = TimedEntry(match, team, scout, board, mutableListOf(), currentTime, { relativeTime })
+        entry = TimedEntry(match, team, scout, board, mutableListOf(), currentTime, { relativeTime.toDouble() })
         startActivityState(WaitingToStart)
     }
