+++ app/src/main/java/ca/warp7/android/scouting/entry/Entry.kt
@@ -68,5 +68,5 @@ interface Entry {
     /**
      * Check if a type should focus according to the current time
      */
-    fun focused(type: Int, time: Int): Boolean
+    fun focused(type: Int, time: Double): Boolean
 }
